Update Docker configuration to use Ubuntu 18.04 as the base image. Install more Perl packages from Ubuntu and less from CPAN. Include sample Apache and SSL configuration files, and include many details about possible local configuration in docker-compose.yml. Some of the run-time configuration is done by setting specific environment variables in docker-compose.yml.
By default the OPL will be installed in a local named Docker volume
when a container is first started, and the OPL-update will be run.
Running OPL-update takes a very long time.
An alternative approach is to modify docker-compose.yml to mount
the OPL from an external directory, which allows for OPL maintenance
to be handled outside of the Docker container/named storage volume.